+++ .dotnet/src/Custom/Assistants/FunctionToolDefinition.cs
@@ -15,7 +15,7 @@ public partial class FunctionToolDefinition : ToolDefinition
     private readonly InternalFunctionDefinition _internalFunction;
 
     /// <inheritdoc cref="InternalFunctionDefinition.Name"/>
-    public required string FunctionName
+    public string FunctionName
     {
         get => _internalFunction.Name;
         set => _internalFunction.Name = value;
@@ -50,15 +50,6 @@ public partial class FunctionToolDefinition : ToolDefinition
     {
         Argument.AssertNotNullOrEmpty(name, nameof(name));
         _internalFunction = new(null, name, null, null, null);
-    }
-
-    /// <summary>
-    /// Creates a new instance of <see cref="FunctionToolDefinition"/>. 
-    /// </summary>
-    public FunctionToolDefinition()
-        : base("function")
-    {
-        _internalFunction = new InternalFunctionDefinition();
     }
 
     [SetsRequiredMembers]