+++ .dotnet/src/Custom/VectorStores/VectorStoreExpirationPolicy.cs
@@ -22,14 +22,14 @@ public partial class VectorStoreExpirationPolicy
     private int _days;
 
     /// <summary> Anchor timestamp after which the expiration policy applies. Supported anchors: `last_active_at`. </summary>
-    public required VectorStoreExpirationAnchor Anchor
+    public VectorStoreExpirationAnchor Anchor
     {
         get => _anchor;
         set => _anchor = value;
     }
 
     /// <summary> The number of days after the anchor time that the vector store will expire. </summary>
-    public required int Days
+    public int Days
     {
         get => _days;
         set => _days = value;
@@ -42,12 +42,6 @@ public partial class VectorStoreExpirationPolicy
     {
         Days = days;
         Anchor = anchor;
-    }
-
-    /// <summary> Initializes a new instance of <see cref="VectorStoreExpirationPolicy"/>. </summary>
-    public VectorStoreExpirationPolicy()
-    {
-        SerializedAdditionalRawData = new ChangeTrackingDictionary<string, BinaryData>();
     }
 
     /// <summary> Initializes a new instance of <see cref="VectorStoreExpirationPolicy"/>. </summary>